TableProvider to separate file
